[dev-vcs/subversion*,www-apache/mod_dav_svn] v. bump to 1.8.8

+++ dev-vcs/subversion/subversion-1.8.8.ebuild
@@ -1,4 +1,4 @@
-# Copyright 1999-2013 Gentoo Foundation
+# Copyright 1999-2014 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
 # $Header: $
 
@@ -6,36 +6,38 @@ EAPI=5
 PYTHON_COMPAT=( python{2_6,2_7} )
 DISTUTILS_OPTIONAL=1
 WANT_AUTOMAKE="none"
+GENTOO_DEPEND_ON_PERL="no"
+
+SAB_PATCHES_SRC=( mirror://sabayon/dev-vcs/${PN}-1.8.5-Gentoo-patches.tar.gz )
+inherit sab-patches autotools bash-completion-r1 db-use depend.apache distutils-r1 elisp-common flag-o-matic libtool multilib perl-module eutils
+
 MY_P="${P/_/-}"
-
-inherit autotools bash-completion-r1 db-use depend.apache distutils-r1 elisp-common flag-o-matic libtool multilib perl-module eutils
-
 DESCRIPTION="Advanced version control system"
 HOMEPAGE="http://subversion.apache.org/"
-SRC_URI="mirror://apache/${PN}/${MY_P}.tar.bz2"
+SRC_URI+=" mirror://apache/${PN}/${MY_P}.tar.bz2"
 S="${WORKDIR}/${MY_P}"
 
 LICENSE="Subversion GPL-2"
 SLOT="0"
 KEYWORDS="~amd64 ~arm ~x86"
-IUSE="apache2 berkdb ctypes-python debug doc +dso extras gnome-keyring java kde nls perl python ruby sasl test vim-syntax +webdav-neon webdav-serf"
+IUSE="apache2 berkdb ctypes-python debug doc +dso extras gnome-keyring +http java kde nls perl python ruby sasl test vim-syntax"
 
-COMMON_DEPEND=">=dev-db/sqlite-3.6.18[threadsafe(+)]
+COMMON_DEPEND=">=dev-db/sqlite-3.7.12
 	>=dev-libs/apr-1.3:1
 	>=dev-libs/apr-util-1.3:1
 	dev-libs/expat
 	sys-libs/zlib
+	app-arch/bzip2
 	berkdb? ( >=sys-libs/db-4.0.14 )
 	ctypes-python? ( ${PYTHON_DEPS} )
 	gnome-keyring? ( dev-libs/glib:2 sys-apps/dbus gnome-base/gnome-keyring )
 	kde? ( sys-apps/dbus dev-qt/qtcore:4 dev-qt/qtdbus:4 dev-qt/qtgui:4 >=kde-base/kdelibs-4:4 )
 	perl? ( dev-lang/perl )
 	python? ( ${PYTHON_DEPS} )
-	ruby? ( >=dev-lang/ruby-1.8.2:1.8
-		dev-ruby/rubygems[ruby_targets_ruby18] )
+	ruby? ( >=dev-lang/ruby-1.9.3:1.9
+		dev-ruby/rubygems[ruby_targets_ruby19] )
 	sasl? ( dev-libs/cyrus-sasl )
-	webdav-neon? ( >=net-libs/neon-0.28 )
-	webdav-serf? ( >=net-libs/serf-0.3.0 )"
+	http? ( >=net-libs/serf-1.2.1 )"
 RDEPEND="${COMMON_DEPEND}
 	apache2? ( www-servers/apache[apache2_modules_dav] )
 	kde? ( kde-base/kwalletd )
@@ -50,7 +52,7 @@ DEPEND="${COMMON_DEPEND}
 	gnome-keyring? ( virtual/pkgconfig )
 	kde? ( virtual/pkgconfig )
 	nls? ( sys-devel/gettext )
-	webdav-neon? ( virtual/pkgconfig )"
+	http? ( virtual/pkgconfig )"
 PDEPEND="java? ( ~dev-vcs/subversion-java-${PV} )"
 
 REQUIRED_USE="
@@ -61,13 +63,13 @@ REQUIRED_USE="
 want_apache
 
 pkg_setup() {
-	if use berkdb; then
+	if use berkdb ; then
 		local apu_bdb_version="$(${EPREFIX}/usr/bin/apu-1-config --includes \
 			| grep -Eoe '-I${EPREFIX}/usr/include/db[[:digit:]]\.[[:digit:]]' \
 			| sed 's:.*b::')"
 		einfo
-		if [[ -z "${SVN_BDB_VERSION}" ]]; then
-			if [[ -n "${apu_bdb_version}" ]]; then
+		if [[ -z "${SVN_BDB_VERSION}" ]] ; then
+			if [[ -n "${apu_bdb_version}" ]] ; then
 				SVN_BDB_VERSION="${apu_bdb_version}"
 				einfo "Matching db version to apr-util"
 			else
@@ -89,27 +91,26 @@ pkg_setup() {
 
 	depend.apache_pkg_setup
 
-	if ! use webdav-neon && ! use webdav-serf; then
+	if ! use http ; then
 		ewarn "WebDAV support is disabled. You need WebDAV to"
 		ewarn "access repositories through the HTTP protocol."
-		ewarn "Consider enabling one of the following USE-flags:"
-		ewarn "  webdav-neon webdav-serf"
+		ewarn "Consider enabling \"http\" USE flag"
 		echo -ne "\a"
 	fi
 
-	if use debug; then
+	if use debug ; then
 		append-cppflags -DSVN_DEBUG -DAP_DEBUG
 	fi
+
+	# http://mail-archives.apache.org/mod_mbox/subversion-dev/201306.mbox/%3C51C42014.3060700@wandisco.com%3E
+	[[ ${CHOST} == *-solaris2* ]] && append-cppflags -D__EXTENSIONS__
 
 	# Allow for custom repository locations.
 	SVN_REPOS_LOC="${SVN_REPOS_LOC:-${EPREFIX}/var/svn}"
 }
 
 src_prepare() {
-	epatch "${FILESDIR}"/${PN}-1.5.4-interix.patch \
-		"${FILESDIR}"/${PN}-1.5.6-aix-dso.patch \
-		"${FILESDIR}"/${PN}-1.6.3-hpux-dso.patch \
-		"${FILESDIR}"/${PN}-fix-parallel-build-support-for-perl-bindings.patch
+	sab-patches_apply_all
 	epatch_user
 
 	fperms +x build/transform_libtool_scripts.sh
@@ -128,7 +129,15 @@ src_prepare() {
 	sed -e 's/\(libsvn_swig_py\)-\(1\.la\)/\1-$(EPYTHON)-\2/g' \
 		-i build-outputs.mk || die "sed failed"
 
-	if use python; then
+	if use python ; then
+		if [[ ${CHOST} == *-darwin* ]] ; then
+			# http://mail-archives.apache.org/mod_mbox/subversion-dev/201306.mbox/%3C20130614113003.GA19257@tarsus.local2%3E
+			# in short, we don't have gnome-keyring stuff here, patch
+			# borrowed from MacPorts
+			die "Darwin not supported; use Gentoo ebuild" # no need to bother with the patch for our needs
+			#epatch "${FILESDIR}"/...swig-python-no-gnome-keyring.patch
+		fi
+
 		# XXX: make python_copy_sources accept path
 		S=${S}/subversion/bindings/swig/python python_copy_sources
 		rm -r "${S}"/subversion/bindings/swig/python || die
@@ -142,12 +151,6 @@ src_configure() {
 		myconf+=" --with-swig"
 	else
 		myconf+=" --without-swig"
-	fi
-
-	if use kde || use nls; then
-		myconf+=" --enable-nls"
-	else
-		myconf+=" --disable-nls"
 	fi
 
 	case ${CHOST} in
@@ -174,9 +177,6 @@ src_configure() {
 		;;
 	esac
 
-	#workaround for bug 387057
-	has_version =dev-vcs/subversion-1.6* && myconf+=" --disable-disallowing-of-undefined-references"
-
 	#version 1.7.7 again tries to link against the older installed version and fails, when trying to
 	#compile for x86 on amd64, so workaround this issue again
 	#check newer versions, if this is still/again needed
@@ -187,11 +187,18 @@ src_configure() {
 		python_export_best
 	fi
 
-	#force ruby-1.8 for bug 399105
-	#allow overriding Python include directory
-	ac_cv_path_RUBY="${EPREFIX}"/usr/bin/ruby18 ac_cv_path_RDOC="${EPREFIX}"/usr/bin/rdoc18 \
+	if use python && [[ ${CHOST} == *-darwin* ]] ; then
+		export ac_cv_python_link="$(tc-getCC) "'$(PYTHON_CFLAGS) -bundle -undefined dynamic_lookup $(PYTHON_LIBS)'
+		export ac_cv_python_libs='$(PYTHON_CFLAGS) -bundle -undefined dynamic_lookup $(PYTHON_LIBS)'
+		export ac_cv_python_compile="$(tc-getCC)"
+	fi
+
+	# force ruby-1.9
+	# allow overriding Python include directory
+	ac_cv_path_RUBY="${EPREFIX}"/usr/bin/ruby19 ac_cv_path_RDOC="${EPREFIX}"/usr/bin/rdoc19 \
 	ac_cv_python_includes='-I$(PYTHON_INCLUDEDIR)' \
 	econf --libdir="${EPREFIX}/usr/$(get_libdir)" \
+		$(use_with apache2 apache-libexecdir) \
 		$(use_with apache2 apxs "${APXS}") \
 		$(use_with berkdb berkeley-db "db.h:${EPREFIX}/usr/include/db${SVN_BDB_VERSION}::db-${SVN_BDB_VERSION}") \
 		$(use_with ctypes-python ctypesgen "${EPREFIX}/usr") \
@@ -199,23 +206,22 @@ src_configure() {
 		$(use_with gnome-keyring) \
 		--disable-javahl \
 		$(use_with kde kwallet) \
+		$(use_enable nls) \
 		$(use_with sasl) \
-		$(use_with webdav-neon neon) \
-		$(use_with webdav-serf serf "${EPREFIX}/usr") \
+		$(use_with http serf) \
 		${myconf} \
 		--with-apr="${EPREFIX}/usr/bin/apr-1-config" \
 		--with-apr-util="${EPREFIX}/usr/bin/apu-1-config" \
 		--disable-experimental-libtool \
 		--without-jikes \
 		--disable-mod-activation \
-		--disable-neon-version-check \
 		--disable-static
 }
 
 src_compile() {
 	emake local-all
 
-	if use ctypes-python; then
+	if use ctypes-python ; then
 		# pre-generate .py files
 		use ctypes-python && emake ctypes-python
 
@@ -224,7 +230,7 @@ src_compile() {
 		popd >/dev/null || die
 	fi
 
-	if use python; then
+	if use python ; then
 		swig_py_compile() {
 			local p=subversion/bindings/swig/python
 			rm -f ${p} || die
@@ -241,27 +247,32 @@ src_compile() {
 		python_foreach_impl swig_py_compile
 	fi
 
-	if use perl; then
+	if use perl ; then
 		emake swig-pl
 	fi
 
-	if use ruby; then
+	if use ruby ; then
 		emake swig-rb
 	fi
 
-	if use extras; then
+	if use extras ; then
 		emake tools
 	fi
 
-	if use doc; then
+	if use doc ; then
 		doxygen doc/doxygen.conf || die "Building of Subversion HTML documentation failed"
 	fi
 }
 
 src_test() {
+	if ! has_version ~${CATEGORY}/${P} ; then
+		ewarn "The test suite shows errors when there is an older version of"
+		ewarn "${CATEGORY}/${PN} installed."
+	fi
+
 	default
 
-	if use ctypes-python; then
+	if use ctypes-python ; then
 		python_test() {
 			"${PYTHON}" subversion/bindings/ctypes-python/test/run_all.py \
 				|| die "ctypes-python tests fail with ${EPYTHON}"
@@ -270,7 +281,7 @@ src_test() {
 		distutils-r1_src_test
 	fi
 
-	if use python; then
+	if use python ; then
 		swig_py_test() {
 			pushd "${BUILD_DIR}" >/dev/null || die
 			"${PYTHON}" tests/run_all.py || die "swig-py tests fail with ${EPYTHON}"
@@ -285,13 +296,13 @@ src_test() {
 src_install() {
 	emake -j1 DESTDIR="${D}" local-install
 
-	if use ctypes-python; then
+	if use ctypes-python ; then
 		pushd subversion/bindings/ctypes-python >/dev/null || die
 		distutils-r1_src_install
 		popd >/dev/null || die
 	fi
 
-	if use python; then
+	if use python ; then
 		swig_py_install() {
 			local p=subversion/bindings/swig/python
 			rm -f ${p} || die
@@ -308,18 +319,18 @@ src_install() {
 		python_foreach_impl swig_py_install
 	fi
 
-	if use perl; then
+	if use perl ; then
 		emake DESTDIR="${D}" INSTALLDIRS="vendor" install-swig-pl
 		fixlocalpod
 		find "${ED}" "(" -name .packlist -o -name "*.bs" ")" -delete
 	fi
 
-	if use ruby; then
+	if use ruby ; then
 		emake DESTDIR="${D}" install-swig-rb
 	fi
 
 	# Install Apache module configuration.
-	if use apache2; then
+	if use apache2 ; then
 		keepdir "${APACHE_MODULES_CONFDIR}"
 		insinto "${APACHE_MODULES_CONFDIR}"
 		doins "${FILESDIR}/47_mod_dav_svn.conf"
@@ -353,14 +364,14 @@ src_install() {
 	rm -fr tools/xslt
 
 	# Install extra files.
-	if use extras; then
+	if use extras ; then
 		cat << EOF > 80subversion-extras
 PATH="${EPREFIX}/usr/$(get_libdir)/subversion/bin"
 ROOTPATH="${EPREFIX}/usr/$(get_libdir)/subversion/bin"
 EOF
 		doenvd 80subversion-extras
 
-		emake DESTDIR="${D}" toolsdir="/usr/$(get_libdir)/subversion/bin" install-tools || die "Installation of tools failed"
+		emake DESTDIR="${D}" toolsdir="/usr/$(get_libdir)/subversion/bin" install-tools
 
 		find tools "(" -name "*.bat" -o -name "*.in" -o -name ".libs" ")" -print0 | xargs -0 rm -fr
 		rm -fr tools/client-side/svnmucc
@@ -372,11 +383,11 @@ EOF
 		doins -r tools
 	fi
 
-	if use doc; then
+	if use doc ; then
 		dohtml -r doc/doxygen/html/*
 	fi
 
-	find "${ED}" '(' -name '*.la' ')' -print0 | xargs -0 rm -f
+	prune_libtool_files --all
 
 	cd "${ED}"usr/share/locale
 	for i in * ; do
@@ -386,10 +397,10 @@ EOF
 
 pkg_preinst() {
 	# Compare versions of Berkeley DB, bug 122877.
-	if use berkdb && [[ -f "${EROOT}usr/bin/svn" ]]; then
+	if use berkdb && [[ -f "${EROOT}usr/bin/svn" ]] ; then
 		OLD_BDB_VERSION="$(scanelf -nq "${EROOT}usr/$(get_libdir)/libsvn_subr-1$(get_libname 0)" | grep -Eo "libdb-[[:digit:]]+\.[[:digit:]]+" | sed -e "s/libdb-\(.*\)/\1/")"
 		NEW_BDB_VERSION="$(scanelf -nq "${ED}usr/$(get_libdir)/libsvn_subr-1$(get_libname 0)" | grep -Eo "libdb-[[:digit:]]+\.[[:digit:]]+" | sed -e "s/libdb-\(.*\)/\1/")"
-		if [[ "${OLD_BDB_VERSION}" != "${NEW_BDB_VERSION}" ]]; then
+		if [[ "${OLD_BDB_VERSION}" != "${NEW_BDB_VERSION}" ]] ; then
 			CHANGED_BDB_VERSION="1"
 		fi
 	fi
@@ -398,7 +409,7 @@ pkg_preinst() {
 pkg_postinst() {
 	use perl && perl-module_pkg_postinst
 
-	if [[ -n "${CHANGED_BDB_VERSION}" ]]; then
+	if [[ -n "${CHANGED_BDB_VERSION}" ]] ; then
 		ewarn "You upgraded from an older version of Berkeley DB and may experience"
 		ewarn "problems with your repository. Run the following commands as root to fix it:"
 		ewarn "    db4_recover -h ${SVN_REPOS_LOC}/repos"
@@ -416,7 +427,7 @@ pkg_config() {
 	# Remember: Don't use ${EROOT}${SVN_REPOS_LOC} since ${SVN_REPOS_LOC}
 	# already has EPREFIX in it
 	einfo "Initializing the database in ${SVN_REPOS_LOC}..."
-	if [[ -e "${SVN_REPOS_LOC}/repos" ]]; then
+	if [[ -e "${SVN_REPOS_LOC}/repos" ]] ; then
 		echo "A Subversion repository already exists and I will not overwrite it."
 		echo "Delete \"${SVN_REPOS_LOC}/repos\" first if you're sure you want to have a clean version."
 	else
@@ -429,7 +440,7 @@ pkg_config() {
 		einfo "Setting repository permissions..."
 		SVNSERVE_USER="$(. "${EROOT}etc/conf.d/svnserve"; echo "${SVNSERVE_USER}")"
 		SVNSERVE_GROUP="$(. "${EROOT}etc/conf.d/svnserve"; echo "${SVNSERVE_GROUP}")"
-		if use apache2; then
+		if use apache2 ; then
 			[[ -z "${SVNSERVE_USER}" ]] && SVNSERVE_USER="apache"
 			[[ -z "${SVNSERVE_GROUP}" ]] && SVNSERVE_GROUP="apache"
 		else
